refactor: MVC 패턴 기반 패키지 분리
