+++ src/main/kotlin/racingcar/domain/Cars.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 class Cars(private val cars: List<Car>) {
     fun move(numberGenerator: NumberGenerator) {
