+++ src/main/kotlin/racingcar/domain/vo/CarAction.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain.vo
 
 enum class CarAction {
     MOVE,
