+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -1,7 +1,7 @@
-package racingcar
+package racingcar.view
 
 class InputView {
-    fun inputCarSize(): String {
+    fun inputCarNames(): String {
         println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).")
         return readln()
     }
